Utilizando fetch para enviar un mail

+++ js/cotizador.js
@@ -18,7 +18,7 @@ const validarFormulario = form.addEventListener('submit', (e) => {
     const descripcion1 = descripcion.value;
     calculador(metrosCuadradosDeseados);
 
-    nombre === "" || email === "" || nroTelefono === "" || ubi === "" || metrosCuadradosDeseados === "" || descripcion1 === "" ? fracaso() : exito (nombre, email, nroTelefono, ubi, metrosCuadradosDeseados, descripcion1, precioTotal);
+    nombre === "" || email === "" || nroTelefono === "" || ubi === "" || metrosCuadradosDeseados === "" || descripcion1 === "" ? fracaso() : exito (nombre, email, nroTelefono, ubi, metrosCuadradosDeseados, descripcion1, precioTotal, e);
 });
 
 let calculador = (metrosCuadradosDeseados) => {
@@ -26,20 +26,80 @@ let calculador = (metrosCuadradosDeseados) => {
     precioTotal = precioMCuadrado*metrosCuadradosDeseados;
 };
 
-let exito = (nombre, email, nroTelefono, ubi, metrosCuadradosDeseados, descripcion1, precioTotal) => {
+let exito = (nombre, email, nroTelefono, ubi, metrosCuadradosDeseados, descripcion1, precioTotal, e) => {
     Swal.fire({
-        icon: "success",
-        title: "Éxito!",
-        text: "El formulario ha sido enviado",
+        title: 'Esta seguro que desea enviar el formulario?',
+        text: `Enviará el formulario`,
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Enviar',
+        cancelButtonText: 'Cancelar'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            Swal.fire(
+                'Enviado!',
+                'El formulario ha sido enviado con éxito',
+                'success'
+            )
+            let informacion = document.getElementById("appl");
+            informacion.innerText = "Hola "+nombre+" la construcción que desea construir en "+ubi+" tiene un valor aproximado entre mano de obra y materiales (sin honorarios) de $"+precioTotal+" para mas información contactarse al 2615788821";
+            sessionStorage.setItem("Nombre", nombre);
+            sessionStorage.setItem("Email", email);
+            sessionStorage.setItem("nroTelefono", nroTelefono);
+            sessionStorage.setItem("Ubicacion", ubi);
+            sessionStorage.setItem("Metros cuadrados", metrosCuadradosDeseados);
+            sessionStorage.setItem("Descripción", descripcion1);
+
+            fetch("https://formsubmit.co/ajax/alvaropereyragamez@gmail.com",{
+                method: "POST",
+                body: new FormData(e.target),
+                })
+                .then(res => res.ok ? res.json() : Promise.reject(res))
+                .then(json => {})
+                .catch (err => {
+                console.log(err);
+                })
+                ;
+
+            Toastify({
+                text: 'Enviaste un mail!',
+                duration: 3000,
+                gravity: 'top',
+                position: 'right',
+                style: {
+                    background: '#4D393B'
+                }
+            }).showToast();
+        }
     })
-    let informacion = document.getElementById("appl");
-    informacion.innerText = "Hola "+nombre+" la construcción que desea construir en "+ubi+" tiene un valor aproximado entre mano de obra y materiales (sin honorarios) de $"+precioTotal+" para mas información contactarse al 2615788821";
-    sessionStorage.setItem("Nombre", nombre);
-    sessionStorage.setItem("Email", email);
-    sessionStorage.setItem("nroTelefono", nroTelefono);
-    sessionStorage.setItem("Ubicacion", ubi);
-    sessionStorage.setItem("Metros cuadrados", metrosCuadradosDeseados);
-    sessionStorage.setItem("Descripción", descripcion1);
+
+    // Swal.fire({
+    //     icon: "success",
+    //     title: "Éxito!",
+    //     text: "El formulario ha sido enviado",
+    // })
+
+    // let informacion = document.getElementById("appl");
+    // informacion.innerText = "Hola "+nombre+" la construcción que desea construir en "+ubi+" tiene un valor aproximado entre mano de obra y materiales (sin honorarios) de $"+precioTotal+" para mas información contactarse al 2615788821";
+    // sessionStorage.setItem("Nombre", nombre);
+    // sessionStorage.setItem("Email", email);
+    // sessionStorage.setItem("nroTelefono", nroTelefono);
+    // sessionStorage.setItem("Ubicacion", ubi);
+    // sessionStorage.setItem("Metros cuadrados", metrosCuadradosDeseados);
+    // sessionStorage.setItem("Descripción", descripcion1);
+
+    // fetch("https://formsubmit.co/ajax/alvaropereyragamez@gmail.com",{
+    //     method: "POST",
+    //     body: new FormData(e.target),
+    //     })
+    //     .then(res => res.ok ? res.json() : Promise.reject(res))
+    //     .then(json => {})
+    //     .catch (err => {
+    //     console.log(err);
+    //     })
+    //     ;
 };
 
 let fracaso = () => {
